(SERVER-1803) Simplify base and uberjar dependency sharing
Previously, a couple of extra variables and a function were used to
control extra dependencies which are excluded from an uberjar build
but which remain present in the dependency info for the project's Maven
artifact.  This commit simplifies that logic a bit more, allowing more
of the dependency info to be listed inline in the defproject macro while
reducing the amount of supporting code for generating the dependency
lists.