(maint) Re-add project.clj at top level in jar
Lein 2.8+ no longer adds `project.clj` as a top-level resource in a jar.
We rely on its being there when we include this library as an additional
uberjar when building pe-puppetserver, so this commits adds project.clj
to the resource list.

Long term it might be better to fix ezbake to handle the new location,
but for the time being, this is the only scenario using that behavior,
so we have decided to fix it here.