(maint) Pin jnr-posix to 3.1.8
JRuby 9.2.17.0 brings in jnr-posix 3.1.4, which is vulnerable to
CVE-2014-4043. jnr-posix 3.1.8 fixes the vulnerability, but JRuby didn't
pick up that version until 9.2.20.0. We can't move to 9.2.20.0 right now
because of a different bug (see SERVER-3133), so for now we are forcing
the upgrade to the first non-vulnerable version.

Note it was necessary to exclude some deps to avoid conflicts with the
versions JRuby brings in.