(maint) Fix CODEOWNERS
This commit updates the CODEOWNERS file to put all teams on the same line.
Previously, by having two separate lines with the `*` rule, the last-defined
rule takes precedence and only the Skeletor team would be applied as owners.
This also updates the comment to remove mention of the old owners.