(SERVER-3167) Take up JRuby 9.4.1.0
This brings in JRuby 9.4.1.0. It also moves us to JRuby's "base" jar
rather than its "core" jar as the core jar is deprecated and superseded
by the base jar.

Puppet Server/jruby-util's appears to interfere with the loading of jars
from within jars installed by gems. This convoluted setup is required
for gems with jar dependencies and with the Ruby standard library being
"gemified" is how JRuby expects Psych/YAML support to load its SnakeYAML
dependency. We did not require this to work before because consumers of
the project already pull in SnakeYAML. In Psych 5.1, on JRuby, it now
depends on SnakeYAML-Engine, which is the successor to SnakeYAML. Puppet
Server and others do not currently pull this in and so we must pull in
SnakeYAML-Engine here.

When Puppet Enterprise takes up this version of JRuby we should evaluate
moving project to use clj-parent and to place SnakeYAML-Engine in
clj-parent.