(maint) Pull joda-time out for individual management
JRuby depends on joda-time and we do some work in the uberjar to exclude
it. What we should really be doing is managing the version in the
uberjar and in normal dependency resolution with clj-parent. However, we
will need to update this libary to use clj-parent when we it up in our
enterprise product. For now simply pull joda-time out into an explicit
dependency.