(maint) Remove defunct uberjar settings
The uberjar settings are quite outdated and are unused by FOSS products.

Instead of trying to fix them for enterprise consumers now we should do
so when this version of JRuby is taken up by our enterprise products and
update the dependency handling for uberjars in the light of using
clj-parent.